Organise files a little better
